Varias mejoras en bbdd, formularios y gestión de errores

+++ lib/providers/database_provider.dart
@@ -1,27 +1,33 @@
-
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:intl/intl.dart';
 
 class DatabaseProvider {
-
   final DateFormat dateFormat = DateFormat('dd/MM/yyyy');
 
   final String? uid;
   DatabaseProvider({this.uid});
 
+  final estadoPublicacion = {
+    'disponible': 'Disponible',
+    'completa': 'Completa',
+    'encurso': 'En curso',
+    'finalizada': 'Finalizada',
+    'cancelada': 'Cancelada',
+  };
+
   //referencias a las colecciones de la base de datos
-  final CollectionReference usuarioCollection = 
-    FirebaseFirestore.instance.collection('usuarios');
+  final CollectionReference usuarioCollection =
+      FirebaseFirestore.instance.collection('usuarios');
 
-  final CollectionReference publicacionesCollection = 
-    FirebaseFirestore.instance.collection('publicaciones');
+  final CollectionReference publicacionesCollection =
+      FirebaseFirestore.instance.collection('publicaciones');
 
-  final CollectionReference reservasCollection = 
-    FirebaseFirestore.instance.collection('reservas');
+  final CollectionReference reservasCollection =
+      FirebaseFirestore.instance.collection('reservas');
 
-  final CollectionReference vehiculosCollection = 
-    FirebaseFirestore.instance.collection('vehiculos');
+  final CollectionReference vehiculosCollection =
+      FirebaseFirestore.instance.collection('vehiculos');
 
   //guardar al usuario en la base de datos cuando se registra
   Future savingUserDataOnRegister(String nombreCompleto, String email) async {
@@ -38,45 +44,64 @@ class DatabaseProvider {
 
   //actualizar nombre de usuario
   Future updateUserName(String nombreCompleto) async {
-    await usuarioCollection.doc(FirebaseAuth.instance.currentUser!.uid)
-    .update({
+    await usuarioCollection.doc(FirebaseAuth.instance.currentUser!.uid).update({
       'nombreCompleto': nombreCompleto,
     });
   }
 
   //buscar al usuario en la base de datos por email
   Future gettingUserDataByEmail(String email) async {
-    QuerySnapshot snapshot = await usuarioCollection.where('email', isEqualTo: email).get();
+    QuerySnapshot snapshot =
+        await usuarioCollection.where('email', isEqualTo: email).get();
     return snapshot;
   }
 
   //get usuario actual
   Future getCurrentUser() async {
-    QuerySnapshot snapshot = await usuarioCollection.where('uid', isEqualTo: FirebaseAuth.instance.currentUser!.uid).get();
+    QuerySnapshot snapshot = await usuarioCollection
+        .where('uid', isEqualTo: FirebaseAuth.instance.currentUser!.uid)
+        .get();
     return snapshot;
   }
 
   //get usuario por uid
   Future getUserByUid(String uid) async {
-    QuerySnapshot snapshot = await usuarioCollection.where('uid', isEqualTo: uid).get();
+    QuerySnapshot snapshot =
+        await usuarioCollection.where('uid', isEqualTo: uid).get();
     return snapshot.docs;
+  }
+
+  //get usuarios que compartan coche en una publicacion
+  Future getUsersSharingCar(String uidPublication) async {
+    QuerySnapshot snapshotReservas = await reservasCollection
+        .where('publicacion', isEqualTo: uidPublication)
+        .get();
+    List<String> usuarios = [];
+    for (var reserva in snapshotReservas.docs) {
+      usuarios.add(reserva['pasajero']);
+    }
+    QuerySnapshot snapshotUsuarios =
+        await usuarioCollection.where('uid', whereIn: usuarios).get();
+    return snapshotUsuarios.docs;
   }
 
   //guardar imagen de perfil
   Future storeProfileImage(String userImage) async {
-    await usuarioCollection.doc(FirebaseAuth.instance.currentUser!.uid)
-    .update({
+    await usuarioCollection.doc(FirebaseAuth.instance.currentUser!.uid).update({
       'imagenPerfil': userImage,
     });
   }
 
   //eliminar usuario de la base de datos
   Future deleteUser() async {
-    await usuarioCollection.doc(FirebaseAuth.instance.currentUser!.uid).delete();
+    await usuarioCollection
+        .doc(FirebaseAuth.instance.currentUser!.uid)
+        .delete();
   }
 
   //guardar vehiculos
-  Future saveVehicle(String matricula, String marca, String modelo, String color) async {
+  Future saveVehicle(
+      String matricula, String marca, String modelo, String color) async {
     final id = vehiculosCollection.doc().id;
     await vehiculosCollection.doc(id).set({
       'matricula': matricula,
@@ -84,7 +109,7 @@ class DatabaseProvider {
       'modelo': modelo,
       'color': color,
       'conductor': FirebaseAuth.instance.currentUser!.uid,
-      'uid' : id,
+      'uid': id,
     });
 
     await usuarioCollection.doc(FirebaseAuth.instance.currentUser!.uid).update({
@@ -94,31 +119,59 @@ class DatabaseProvider {
 
   //get vehiculo por uid
   Future getVehicleByUid(String uid) async {
-    QuerySnapshot snapshot = await vehiculosCollection.where('uid', isEqualTo: uid).get();
+    QuerySnapshot snapshot =
+        await vehiculosCollection.where('uid', isEqualTo: uid).get();
     return snapshot.docs;
   }
-  
+
   //get vehiculos de usuario
   Future getVehicles() async {
-    QuerySnapshot snapshot = await vehiculosCollection.where('conductor', isEqualTo: FirebaseAuth.instance.currentUser!.uid).get();
+    QuerySnapshot snapshot = await vehiculosCollection
+        .where('conductor', isEqualTo: FirebaseAuth.instance.currentUser!.uid)
+        .get();
     return snapshot.docs;
   }
 
   //eliminar vehiculo
   Future deleteVehicle(String uid) async {
-    await vehiculosCollection.doc(uid).delete();
-    await usuarioCollection.doc(FirebaseAuth.instance.currentUser!.uid).update({
-      'vehiculos': FieldValue.arrayRemove([uid]),
-    });
+    //get publicaciones del vehiculo
+    QuerySnapshot snapshotPublicaciones =
+        await publicacionesCollection.where('vehiculo', isEqualTo: uid).get();
+
+    if (snapshotPublicaciones.docs.isNotEmpty) {
+      return false;
+    } else {
+      await vehiculosCollection.doc(uid).delete();
+      await usuarioCollection
+          .doc(FirebaseAuth.instance.currentUser!.uid)
+          .update({
+        'vehiculos': FieldValue.arrayRemove([uid]),
+      });
+      return true;
+    }
   }
 
   //guardar publicacion
-  Future savePublication(int fecha, String hora, String origen, String destino, String coordenadasOrigen, String coordenadasDestino, int asientos, double precio, String uidVehiculo) async {
+  Future savePublication(
+    int fecha,
+    String duracionViaje,
+    String distancia,
+    String hora,
+    String origen,
+    String destino,
+    String coordenadasOrigen,
+    String coordenadasDestino,
+    int asientos,
+    double precio,
+    String uidVehiculo,
+  ) async {
     final id = vehiculosCollection.doc().id;
     await publicacionesCollection.doc(id).set({
       'pasajeros': [],
       'fecha': fecha,
       'horaSalida': hora,
+      'duracionViaje': duracionViaje,
+      'distancia': distancia,
       'origen': origen,
       'destino': destino,
       'coordenadasOrigen': coordenadasOrigen,
@@ -127,6 +180,7 @@ class DatabaseProvider {
       'precio': precio,
       'vehiculo': uidVehiculo,
       'conductor': FirebaseAuth.instance.currentUser!.uid,
+      'estado': estadoPublicacion['disponible'],
       'uid': id,
     });
 
@@ -135,28 +189,37 @@ class DatabaseProvider {
     });
   }
 
-  //add pasajero a la publicacion
-  Future addPassengerToPublication(String uidPublicacion, String uidPasajero) async {
-    await publicacionesCollection.doc(uidPublicacion).update({
-      'pasajeros': FieldValue.arrayUnion([uidPasajero]),
-    });
-  }
-
   //get publicaciones
   Future getPublications(String id) async {
-    QuerySnapshot snapshot = await publicacionesCollection.where('uid', isEqualTo: id).get();
+    QuerySnapshot snapshot =
+        await publicacionesCollection.where('uid', isEqualTo: id).get();
     return snapshot.docs;
   }
 
   //get publicaciones hacia el Zonzamas
   Future getPublicationsToZonzamas([dynamic fecha]) async {
     if (fecha != null) {
-      QuerySnapshot snapshot = await publicacionesCollection.where('fecha', isEqualTo: fecha).where('destino', isEqualTo: 'CIFP Zonzamas').orderBy('horaSalida').orderBy('origen').get();
+      QuerySnapshot snapshot = await publicacionesCollection
+          .where('fecha', isGreaterThanOrEqualTo: fecha)
+          .where('fecha', isLessThanOrEqualTo: fecha + 86399000)
+          .where('destino', isEqualTo: 'CIFP Zonzamas')
+          .where('estado', isEqualTo: estadoPublicacion['disponible'])
+          .orderBy('fecha')
+          .orderBy('horaSalida')
+          .orderBy('origen')
+          .get();
       return snapshot.docs;
     } else {
       DateTime today = DateTime.now();
       final dateToSearch = today.millisecondsSinceEpoch;
-      QuerySnapshot snapshot = await publicacionesCollection.where('fecha', isGreaterThanOrEqualTo: dateToSearch).where('destino', isEqualTo: 'CIFP Zonzamas').orderBy('fecha').orderBy('horaSalida').orderBy('origen').get();
+      QuerySnapshot snapshot = await publicacionesCollection
+          .where('fecha', isGreaterThanOrEqualTo: dateToSearch)
+          .where('destino', isEqualTo: 'CIFP Zonzamas')
+          .where('estado', isEqualTo: estadoPublicacion['disponible'])
+          .orderBy('fecha')
+          .orderBy('horaSalida')
+          .orderBy('origen')
+          .get();
       return snapshot.docs;
     }
   }
@@ -164,23 +227,62 @@ class DatabaseProvider {
   //get publicaciones desde el Zonzamas
   Future getPublicationsFromZonzamas([dynamic fecha]) async {
     if (fecha != null) {
-      
-      QuerySnapshot snapshot = await publicacionesCollection.where('fecha', isEqualTo: fecha).where('origen', isEqualTo: 'CIFP Zonzamas').orderBy('horaSalida').orderBy('destino').get();
+      QuerySnapshot snapshot = await publicacionesCollection
+          .where('fecha', isGreaterThanOrEqualTo: fecha)
+          .where('fecha', isLessThanOrEqualTo: fecha + 86399000)
+          .where('origen', isEqualTo: 'CIFP Zonzamas')
+          .where('estado', isEqualTo: estadoPublicacion['disponible'])
+          .orderBy('fecha')
+          .orderBy('horaSalida')
+          .orderBy('destino')
+          .get();
       return snapshot.docs;
     } else {
       DateTime today = DateTime.now();
       final dateToSearch = today.millisecondsSinceEpoch;
-      QuerySnapshot snapshot = await publicacionesCollection.where('fecha', isGreaterThanOrEqualTo: dateToSearch).where('origen', isEqualTo: 'CIFP Zonzamas').orderBy('fecha').orderBy('horaSalida').orderBy('destino').get();
+      QuerySnapshot snapshot = await publicacionesCollection
+          .where('fecha', isGreaterThanOrEqualTo: dateToSearch)
+          .where('origen', isEqualTo: 'CIFP Zonzamas')
+          .where('estado', isEqualTo: estadoPublicacion['disponible'])
+          .orderBy('fecha')
+          .orderBy('horaSalida')
+          .orderBy('destino')
+          .get();
       return snapshot.docs;
     }
   }
 
-  //get publicaciones de usuario
-  Future getPublicationsByUser() async {
-    QuerySnapshot snapshot = await publicacionesCollection.where('conductor', isEqualTo: FirebaseAuth.instance.currentUser!.uid).get();
+  //get publicaciones dónde el usuario es conductor
+  Future getPublicationsByDriver() async {
+    QuerySnapshot snapshot = await publicacionesCollection
+        .where('conductor', isEqualTo: FirebaseAuth.instance.currentUser!.uid)
+        .get();
     return snapshot.docs;
   }
-  
+
+  //get publicaciones dónde el usuario es pasajero
+  Future getPublicationsByPassenger() async {
+    QuerySnapshot snapshot = await publicacionesCollection
+        .where(
+          'pasajeros',
+          arrayContains: FirebaseAuth.instance.currentUser!.uid,
+        )
+        .get();
+    return snapshot.docs;
+  }
+
+  //comprobar que el viaje no está lleno
+  Future<bool> checkIfFull(String uid) async {
+    DocumentSnapshot snapshot = await publicacionesCollection.doc(uid).get();
+    final asientosDisponibles = snapshot['asientosDisponibles'];
+    final pasajeros = snapshot['pasajeros'];
+    if (asientosDisponibles > pasajeros.length) {
+      return false;
+    } else {
+      return true;
+    }
+  }
+
   //guardar reserva
   Future saveReservation(String uidPublicacion, String uidPasajero) async {
     final id = reservasCollection.doc().id;
@@ -196,15 +298,12 @@ class DatabaseProvider {
 
     await publicacionesCollection.doc(uidPublicacion).update({
       'pasajeros': FieldValue.arrayUnion([uidPasajero]),
-      'plazasDisponibles': FieldValue.increment(-1),
     });
-  }
 
-  //get reservas de usuario
-  Future getReservationsByUser() async {
-    QuerySnapshot snapshot = await reservasCollection.where('pasajero', isEqualTo: FirebaseAuth.instance.currentUser!.uid).get();
-    return snapshot.docs;
+    if (await checkIfFull(uidPublicacion)) {
+      await publicacionesCollection.doc(uidPublicacion).update({
+        'estado': estadoPublicacion['completa'],
+      });
+    }
   }
-
-  
 }
